set text size

+++ src/maximsblog/blogspot/com/jlatexmath/ExampleFragment.java
@@ -20,10 +20,12 @@ import android.graphics.RectF;
 import android.os.Bundle;
 import android.os.Environment;
 import android.support.v4.app.Fragment;
+import android.util.TypedValue;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.View.OnClickListener;
 import android.view.ViewGroup;
+import android.widget.EditText;
 import android.widget.ImageView;
 import android.widget.LinearLayout;
 import android.widget.RelativeLayout;
@@ -31,28 +33,24 @@ import android.widget.Toast;
 
 public class ExampleFragment extends Fragment implements OnClickListener {
 
-	public static android.support.v4.app.Fragment newInstance(String latex,
-			int size, int tag) {
+	public static android.support.v4.app.Fragment newInstance(String latex, int tag) {
 		ExampleFragment fragment = new ExampleFragment();
 		fragment.setTag(tag);
 		fragment.setFormula(latex);
-		fragment.setTextSize(size);
 		return fragment;
 	}
 
 	private ImageView mImageView;
 	Bitmap mImage;
 	private String mLatex;
-	private float mTextSize = 16;
+	private float mTextSize = 10;
 	private int mTag;
+	private EditText mSizeText;
 
 	private void setFormula(String latex) {
 		mLatex = latex;
 	}
 
-	private void setTextSize(float size) {
-		mTextSize = size;
-	}
 
 	private void setTag(int tag) {
 		mTag = tag;
@@ -79,16 +77,25 @@ public class ExampleFragment extends Fragment implements OnClickListener {
 	@Override
 	public View onCreateView(LayoutInflater inflater, ViewGroup container,
 			Bundle savedInstanceState) {
-
+		
 		LinearLayout layout = (LinearLayout) inflater.inflate(
 				R.layout.fragment_example, container, false);
 		mImageView = (ImageView) layout.findViewById(R.id.imageView1);
+		mSizeText = (EditText) layout.findViewById(R.id.size);
+		layout.findViewById(R.id.button1).setOnClickListener(this);
 		layout.findViewById(R.id.save).setOnClickListener(this);
+		setformula();
 
+		return layout;
+	}
+
+	private void setformula() {
+		int w = getResources().getDisplayMetrics().widthPixels;
+		int h = getResources().getDisplayMetrics().heightPixels;
 		TeXFormula formula = new TeXFormula(mLatex);
-
 		TeXIcon icon = formula.new TeXIconBuilder()
 				.setStyle(TeXConstants.STYLE_DISPLAY).setSize(mTextSize)
+				.setWidth(TeXConstants.UNIT_CC, 1, TeXConstants.ALIGN_LEFT)
 				.build();
 		icon.setInsets(new Insets(5, 5, 5, 5));
 
@@ -98,57 +105,61 @@ public class ExampleFragment extends Fragment implements OnClickListener {
 		Canvas g2 = new Canvas(mImage);
 		g2.drawColor(Color.WHITE);
 		icon.paintIcon(g2, 0, 0);
-		int w = getResources().getDisplayMetrics().widthPixels;
-		int h = getResources().getDisplayMetrics().heightPixels;
 
 		Bitmap scaleimage = scaleBitmapAndKeepRation(mImage, h, w);
-		mImage.recycle();
+
 		mImageView.setImageBitmap(scaleimage);
-		/**/
-
-		return layout;
 	}
-
-	public static Bitmap scaleBitmapAndKeepRation(Bitmap TargetBmp,
+	
+	public Bitmap scaleBitmapAndKeepRation(Bitmap targetBmp,
 			int reqHeightInPixels, int reqWidthInPixels) {
-		Matrix m = new Matrix();
-		m.setRectToRect(
-				new RectF(0, 0, TargetBmp.getWidth(), TargetBmp.getHeight()),
-				new RectF(0, 0, reqWidthInPixels, reqHeightInPixels),
-				Matrix.ScaleToFit.CENTER);
-		Bitmap scaledBitmap = Bitmap.createBitmap(TargetBmp, 0, 0,
-				TargetBmp.getWidth(), TargetBmp.getHeight(), m, true);
-		return scaledBitmap;
+		/*
+		 * if (targetBmp.getWidth() > reqWidthInPixels || targetBmp.getHeight()
+		 * > reqHeightInPixels) { Matrix m = new Matrix(); m.setRectToRect( new
+		 * RectF(0, 0, targetBmp.getWidth(), targetBmp.getHeight()), new
+		 * RectF(0, 0, reqWidthInPixels, reqHeightInPixels),
+		 * Matrix.ScaleToFit.CENTER); Bitmap scaledBitmap =
+		 * Bitmap.createBitmap(targetBmp, 0, 0, targetBmp.getWidth(),
+		 * targetBmp.getHeight(), m, true); mImage.recycle(); return
+		 * scaledBitmap; }
+		 */
+		return targetBmp;
 	}
+
 	@Override
 	public void onDestroy() {
-		//mImage.recycle();
 		super.onDestroy();
 	};
 
 	@Override
 	public void onClick(View v) {
-		File dir = getActivity().getExternalCacheDir();
-		String r;
-		if (dir != null) {
-			dir.mkdirs();
-			String fname = "Example" + (mTag + 1) + ".png";
-			File file = new File(dir, fname);
-			if (file.exists())
-				file.delete();
+		if (v.getId() == R.id.button1) {
+			mTextSize = Integer.valueOf(mSizeText.getText().toString());
+			setformula();
+		} else {
+			File dir = getActivity().getExternalCacheDir();
+			String r;
+			if (dir != null) {
+				dir.mkdirs();
+				String fname = "Example" + (mTag + 1) + ".png";
+				File file = new File(dir, fname);
+				if (file.exists())
+					file.delete();
 
-			try {
-				FileOutputStream out = new FileOutputStream(file);
-				mImage.compress(Bitmap.CompressFormat.PNG, 90, out);
-				out.flush();
-				out.close();
-				r = getString(R.string.saved) + ' ' + file.getAbsolutePath();
-			} catch (Exception e) {
-				e.printStackTrace();
-				r = getString(R.string.err) + ' ' + e.toString();
-			}
-		} else
-			r = getString(R.string.err);
-		Toast.makeText(getActivity(), r, Toast.LENGTH_LONG).show();
+				try {
+					FileOutputStream out = new FileOutputStream(file);
+					mImage.compress(Bitmap.CompressFormat.PNG, 90, out);
+					out.flush();
+					out.close();
+					r = getString(R.string.saved) + ' '
+							+ file.getAbsolutePath();
+				} catch (Exception e) {
+					e.printStackTrace();
+					r = getString(R.string.err) + ' ' + e.toString();
+				}
+			} else
+				r = getString(R.string.err);
+			Toast.makeText(getActivity(), r, Toast.LENGTH_LONG).show();
+		}
 	}
 }
